Added utils.py
+ added utils.py file
~ Moved image_tag, image_id fixtures to utils.py
~ Updated tests and pre-commit hooks for importing fixtures
