+++ tox.ini
@@ -1,4 +1,6 @@
 [flake8]
 per-file-ignores =
-    __init__.py: F401
+    __init__.py: F401,
+    test/test_image_internals.py: F401, F811,
+    test/test_image.py: F401, F811
 max-line-length = 88
